+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- support configuring Git executable path
+
 ## [13.0.5-dev.9 - 2018-10-03]
 
 - support print multiple external (build / test) errors